Renamed teamMembers file.
